ENGR00223120 Ethernet:fix an priority issue in concurrent
As default, ethernet is not added in framework and connectivityservice
also not consider it, when adding it in config.xml,there will be a "mobile,
wifi, ethernet" concurrent case falling into forever loopback.
For only two types of network,the logic has no issue in tryFailover(),but
if three types it will has this issue, so priority is needed to optimise the
reconnect process.

Signed-off-by: Jianzheng Zhou <B38613@freescale.com>